Send money to friends

+++ app/controllers/gifts_controller.rb
@@ -2,8 +2,29 @@ class GiftsController < ApplicationController
 
   def index
     @cards = current_user.cards.select do |card|
-      card.color == "green" && !card.played?
+      card.game.color == "green" && !card.played?
     end
+  end
+  
+  def card
+  end
+  
+  def send_card
+    card = Card.find(params[:card_id])
+    card.user = nil
+    card.gift_for = params[:id].to_i
+    card.save
+    redirect_to(gifts_url)
+  end
+  
+  def receive_card
+    card = Card.find(params[:id])
+    if card.gift_for == current_user.facebook_id
+      card.user = current_user.id
+      card.gift_for = nil
+      card.save
+    end
+    redirect_to(cards_url)
   end
 
   def money
@@ -19,8 +40,8 @@ class GiftsController < ApplicationController
   end
 
   def receive_money
-    m = MoneyGift.find(params[:id])
-    if current_user.facebook_id == m.facebook_id
+    m = MoneyGift.find(:first, :conditions => {:facebook_id => current_user.facebook_id})
+    if !m.nil? && current_user.facebook_id == m.facebook_id
       current_user.money += m.value
       current_user.save
       m.destroy
